analyze_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
Pillow>=11.0.0
transformers>=4.38.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.7.0
scipy>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
